backend/requirements.txt
django==3.0.5
djangorestframework 
pymongo==3.12.3
Pytz
# sqlparse==0.3.1
sqlparse==0.2.4
axios